requirements.txt
Django==4.2.27
Pillow==11.1.0
django-unfold==0.45.0
